Extract FakeAuthService from SearchServiceTest
