+++ app/src/test/java/net/squanchy/search/SearchServiceTest.kt
@@ -1,18 +1,14 @@
 package net.squanchy.search
 
-import arrow.core.Option
-import com.google.android.gms.auth.api.signin.GoogleSignInAccount
-import io.reactivex.Completable
 import io.reactivex.Observable
+import net.squanchy.FakeAuthService
 import net.squanchy.schedule.domain.view.anEvent
-import net.squanchy.search.algolia.AlgoliaSearchEngine
-import net.squanchy.search.algolia.model.AlgoliaSearchResult
 import net.squanchy.search.SearchListElement.EventElement
 import net.squanchy.search.SearchListElement.SpeakerElement
-import net.squanchy.service.repository.AuthService
+import net.squanchy.search.algolia.AlgoliaSearchEngine
+import net.squanchy.search.algolia.model.AlgoliaSearchResult
 import net.squanchy.service.repository.EventRepository
 import net.squanchy.service.repository.SpeakerRepository
-import net.squanchy.service.repository.User
 import net.squanchy.speaker.domain.view.aSpeaker
 import org.junit.Before
 import org.junit.Rule
@@ -28,20 +24,20 @@ class SearchServiceTest {
     @JvmField
     var rule: MockitoRule = MockitoJUnit.rule()
 
-    lateinit var searchService: SearchService
+    private lateinit var searchService: SearchService
 
     @Mock
-    lateinit var eventRepository: EventRepository
+    private lateinit var eventRepository: EventRepository
 
     @Mock
-    lateinit var speakerRepository: SpeakerRepository
+    private lateinit var speakerRepository: SpeakerRepository
 
     @Mock
-    lateinit var algoliaSearchEngine: AlgoliaSearchEngine
+    private lateinit var algoliaSearchEngine: AlgoliaSearchEngine
 
     @Before
     fun setup() {
-        searchService = SearchService(eventRepository, speakerRepository, FakeAuthService, algoliaSearchEngine)
+        searchService = SearchService(eventRepository, speakerRepository, FakeAuthService(UID), algoliaSearchEngine)
     }
 
     @Test
@@ -98,32 +94,5 @@ class SearchServiceTest {
     companion object {
         private const val QUERY = "A"
         private const val UID = "uid"
-    }
-
-    private object FakeAuthService : AuthService {
-
-        override fun signInWithGoogle(account: GoogleSignInAccount): Completable {
-            TODO("not implemented")
-        }
-
-        override fun <T> ifUserSignedInThenObservableFrom(observable: (String) -> Observable<T>): Observable<T> {
-            return observable(UID)
-        }
-
-        override fun ifUserSignedInThenCompletableFrom(completable: (String) -> Completable): Completable {
-            TODO("not implemented")
-        }
-
-        override fun currentUser(): Observable<Option<User>> {
-            TODO("not implemented")
-        }
-
-        override fun signOut(): Completable {
-            TODO("not implemented")
-        }
-
-        override fun signInAnonymously(): Completable {
-            TODO("not implemented")
-        }
     }
 }
